+++ boards/board-arduino-uno/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @typecad/board-arduino-uno
 
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- @typecad/cuttlefish@1.0.0-alpha.8
+- @typecad/hal@1.0.0-alpha.8
+- @typecad/mcu-atmega328p@1.0.0-alpha.8
+
 ## 1.0.0-alpha.7
 
 ### Patch Changes